Mostly finished customizing bash prompt.
This commit mostly finishes the ability to completely customize the bash
prompt in a manner similar to __git_ps1. In fact, it replaces
PROMPT_COMMAND with __bash_ps1 vs. __git_ps1 and is fully functional.

This commit contains lots of commented out echo statements used for
debugging, however. These will be removed in a future commit.