Update README and some variable renaming
This commit updates the documentation provided in the README.md file.
While writing this documentation, I realized I could simplify some
variable names, as well as the way I handled whether or not a separator
would be show between ahead/behind changes in the git format string when a
branch is in a diverged state and GIT_PS1_SHOWUPSTREAM='auto'.