Some README.md formatting updates.
This commit adds a lot of HTML markup to the README.md because GitHub
doesn't currently support the def_list Pagedown markdown extension. It
also doesn't support the very useful StackEdit.io markdown [TOC] which
creates a table of contens out of all of the document markdown headers.

If GitHub at some point supports the def_list module, I'll remove a lot of
the HTML and replace it with the appropriate markdown. I put a feature
request into GitHub for them to support the def_list extension. For more
info on this markdown extension, see
https://github.com/jmcmanus/pagedown-extra.