Fix .bashrc PROMPT_COMMAND value.
A couple of PROMPT_COMMAND value fixes so that when this file is retrieved
from the repository, it can be used as-is and it will cause bash to
display a standard shell prompt with a standard Git repository format
string.

Also fixed sourcing sh-prompt.sh from .bashrc (since it was renamed from
bash-prompt.sh).