Simplified the SH_FORMAT_STRING formatter.
Before, this variable expected various tokens such as %u and %h for
representing the typical parts of a shell prompt (i.e. username and
hostname). But, the '%' is not really needed since the value of this
variable should not contain anything but a combination of one or more of
the letters u, z, h, w, and v--and nothing else.

In fact,  getting rid of the '%' symbol should allow me to simplify this
script a bit more, which will hopefully result in speed improvements.

NOTE: At this point, the prompt is maybe as much as .25s slower than just
using __git_ps1 alone. So that's not too bad. But I'd like it to be
snappier yet.