Replace if statements with case. Fix arg processing
This commit fixes how arguments are processed. It was possible with the
previous implementation that the arguments could have been processed
incorrectly.

More importantly, this commit removes most instances of 'if' in favor of
'case' statements. Wow! You really would be surprised at how mach slower
'if' constructs are over 'case'.

Now sh-prompt.sh is really fast. The prompt only becomes slow when you're
in a git repository. Part of that is because of git-prompt.sh itself, but
I do need to see if any additions I made may be causing slow-downs (e.g.
more 'if's where 'case's would do).