Add the ability to change git dirty state display
This commit adds the aility to set the environment variable
GIT_PS1_DIRTYSTATE_FORMAT in your .bashrc file to specify the order in
which dirty state information for a branch should be displayed.

By default, the value of this variable if undeclared, or empty is 'wisu',
which results in dirty state information being displayed in the following
order: unstaged changes (*), staged changes (+), stashed changesets ($),
followed by untracked files (%). Note that this is the standard
git-prompt.sh default.

Of course, the strings (and their color) can be customized with other
environment variables. Now you can change the order in which this
information is displayed.