Add ability to always show branch state info
This commit gives you the ability to always show branch/repository state
indicators. The intent is to use this option to provide "watermark"
indicators, or indicators that appear "dimmed" when th branch is clean.
Then when something interesting occurs, the appropriate indicator "lights"
up. I don't like things moving around too much (e.g. one moment the prompt
displays %, then *%, the +%, then +*%-- I'd rather just always see +*%$ in
a "grayed out" state, and when the branch becomes "dirty", the indicator
lights up)

With this commit, there is only one color for the clean state; so it's all
or nothing. A future commit will add the ability to provide a "clean"
color for each of the indicators to provide a true "light up" effect.